# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Retriva Gateway API
  description: |
    Retriva Gateway acts as the Backend-for-Frontend (BFF) for the Retriva WebUI
    and provides a public-facing API for external integrations.

    It exposes:
    * `/gateway/*`: Backend-for-Frontend (BFF) routes specifically designed for the WebUI.
    * `/api/v2/*`: Public-facing modern API routes.
  version: 1.0.0
  contact:
    name: Andrea Marson
    email: am.dev.75@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8002
    description: Retriva Gateway API
...
paths:
...
  /api/v2/sources:
    post:
      tags: [Connected Sources v2]
      summary: Create a connected source
      operationId: createSourceV2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceRequest"
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "422":
          description: Invalid connector type or config
        "503":
          description: Dynamic ingestion disabled
    get:
      tags: [Connected Sources v2]
      summary: List connected sources
      operationId: listSourcesV2
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Source list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceResponse"
  /api/v2/sources/{source_id}:
    get:
      tags: [Connected Sources v2]
      summary: Get source details
      operationId: getSourceV2
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Source details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "404":
          description: Source not found
    patch:
      tags: [Connected Sources v2]
      summary: Update source
      operationId: updateSourceV2
      parameters:
        - $ref: "#/components/parameters/SourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSourceRequest"
      responses:
        "200":
          description: Source updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
    delete:
      tags: [Connected Sources v2]
      summary: Delete source
      operationId: deleteSourceV2
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "204":
          description: Source deleted
  /api/v2/sources/{source_id}/sync:
    post:
      tags: [Connected Sources v2]
      summary: Trigger manual sync
      operationId: triggerSyncV2
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Sync run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceRunResponse"
        "409":
          description: Source in invalid state for sync
  /api/v2/sources/{source_id}/pause:
    post:
      tags: [Connected Sources v2]
      summary: Pause source
      operationId: pauseSourceV2
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Source paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "409":
          description: Source cannot be paused from current state
  /api/v2/sources/{source_id}/resume:
    post:
      tags: [Connected Sources v2]
      summary: Resume source
      operationId: resumeSourceV2
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Source resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "409":
          description: Source is not paused
  /api/v2/sources/{source_id}/status:
    get:
      tags: [Connected Sources v2]
      summary: Get source status
      operationId: getSourceStatusV2
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Source status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceStatusResponse"
  /api/v2/sources/{source_id}/runs:
    get:
      tags: [Connected Sources v2]
      summary: List sync runs
      operationId: listRunsV2
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Runs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceRunResponse"
  /api/v2/sources/{source_id}/runs/{run_id}:
    get:
      tags: [Connected Sources v2]
      summary: Get run details
      operationId: getRunV2
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceRunResponse"
        "404":
          description: Run not found
  /api/v2/sources/{source_id}/items:
    get:
      tags: [Connected Sources v2]
      summary: List synced items
      operationId: listItemsV2
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceItemState"

  # --- Connected Sources (Dynamic Ingestion) ---
  /gateway/sources:
    post:
      tags: [Connected Sources]
      summary: Create a connected source
      operationId: createSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceRequest"
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "422":
          description: Invalid connector type or config
        "503":
          description: Dynamic ingestion disabled
    get:
      tags: [Connected Sources]
      summary: List connected sources
      operationId: listSources
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Source list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceResponse"
  /gateway/sources/{source_id}:
    get:
      tags: [Connected Sources]
      summary: Get source details
      operationId: getSource
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Source details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "404":
          description: Source not found
    patch:
      tags: [Connected Sources]
      summary: Update source
      operationId: updateSource
      parameters:
        - $ref: "#/components/parameters/SourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSourceRequest"
      responses:
        "200":
          description: Source updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
    delete:
      tags: [Connected Sources]
      summary: Delete source
      operationId: deleteSource
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "204":
          description: Source deleted
  /gateway/sources/{source_id}/sync:
    post:
      tags: [Connected Sources]
      summary: Trigger manual sync
      operationId: triggerSync
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Sync run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceRunResponse"
        "409":
          description: Source in invalid state for sync
  /gateway/sources/{source_id}/pause:
    post:
      tags: [Connected Sources]
      summary: Pause source
      operationId: pauseSource
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Source paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "409":
          description: Source cannot be paused from current state
  /gateway/sources/{source_id}/resume:
    post:
      tags: [Connected Sources]
      summary: Resume source
      operationId: resumeSource
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Source resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "409":
          description: Source is not paused
  /gateway/sources/{source_id}/status:
    get:
      tags: [Connected Sources]
      summary: Get source status
      operationId: getSourceStatus
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Source status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceStatusResponse"
  /gateway/sources/{source_id}/runs:
    get:
      tags: [Connected Sources]
      summary: List sync runs
      operationId: listRuns
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Runs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceRunResponse"
  /gateway/sources/{source_id}/runs/{run_id}:
    get:
      tags: [Connected Sources]
      summary: Get run details
      operationId: getRun
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceRunResponse"
        "404":
          description: Run not found
  /gateway/sources/{source_id}/items:
    get:
      tags: [Connected Sources]
      summary: List synced items
      operationId: listItems
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceItemState"

  # --- Internal Worker API ---
...
components:
  schemas:
...
    SourceStatus:
      type: string
      enum: [CREATED, VALIDATING_CONNECTION, BASELINE_PENDING, BASELINE_RUNNING, CATCHUP_RUNNING, ACTIVE, PAUSED, DEGRADED, FAILED, DELETING, DELETED]
    SyncMode:
      type: string
      enum: [baseline, catchup, incremental]
...
    RunPhase:
      type: string
      enum: [baseline_scan, catchup, incremental, validation]
    RunStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
    ItemStatus:
      type: string
      enum: [pending, indexed, deleted, error]
    CreateSourceRequest:
      type: object
      required: [connector_type, display_name, target_kb_id]
      properties:
        connector_type: { type: string }
        display_name: { type: string }
        target_kb_id: { type: string }
        config: { type: object, additionalProperties: {} }
        secret_ref: { type: string, nullable: true }
        metadata: { type: object, additionalProperties: {} }
        schedule: { type: string, nullable: true }
        idempotency_key: { type: string, nullable: true }
    UpdateSourceRequest:
      type: object
      properties:
        display_name: { type: string }
        config: { type: object, additionalProperties: {} }
        schedule: { type: string }
        metadata: { type: object, additionalProperties: {} }
    SourceResponse:
      type: object
      required: [source_id, tenant_id, connector_type, display_name, target_kb_id, status, sync_mode, created_at, updated_at]
      properties:
        source_id: { type: string }
        tenant_id: { type: string }
        connector_type: { type: string }
        display_name: { type: string }
        target_kb_id: { type: string }
        status:
          $ref: "#/components/schemas/SourceStatus"
        sync_mode:
          $ref: "#/components/schemas/SyncMode"
        schedule: { type: string, nullable: true }
        config: { type: object, additionalProperties: {} }
        has_secret: { type: boolean }
        metadata: { type: object, additionalProperties: {} }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SourceRunResponse:
      type: object
      required: [run_id, source_id, phase, status, started_at]
      properties:
        run_id: { type: string }
        source_id: { type: string }
        phase:
          $ref: "#/components/schemas/RunPhase"
        status:
          $ref: "#/components/schemas/RunStatus"
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
        processed_items: { type: integer, default: 0 }
        failed_items: { type: integer, default: 0 }
        skipped_items: { type: integer, default: 0 }
        error_code: { type: string, nullable: true }
    SourceStatusResponse:
      type: object
      required: [source_id, status, sync_mode]
      properties:
        source_id: { type: string }
        status:
          $ref: "#/components/schemas/SourceStatus"
        sync_mode:
          $ref: "#/components/schemas/SyncMode"
        latest_run:
          $ref: "#/components/schemas/SourceRunResponse"
        has_checkpoint: { type: boolean }
    SourceItemState:
      type: object
      required: [source_id, source_item_id]
      properties:
        source_id: { type: string }
        source_item_id: { type: string }
        source_revision: { type: string, nullable: true }
        retriva_doc_id: { type: string, nullable: true }
        content_hash: { type: string, nullable: true }
        status:
          $ref: "#/components/schemas/ItemStatus"
        last_synced_at: { type: string, format: date-time }
...
  parameters:
    SourceId:
      name: source_id
      in: path
      required: true
      schema: { type: string }
      description: Connected source ID (e.g. src_a1b2c3d4e5f6)
    RunId:
      name: run_id
      in: path
      required: true
      schema: { type: string }
      description: Sync run ID (e.g. run_a1b2c3d4e5f6)
...